Initial commit of docker component charm
this charm is intended to be a base, for use in juju-compose layering.
This particular charm reduces the overall footprint of the base charm
from a 7 minute installation path, to a 2 minute installation path
(including provider VM/metal spinup time)

THere are no relations, and only a configured docker user. Every
additional layer on top of this base can and should extend the charm
into a useful state to be shared in the ecosystem of charming.